requirements.txt
alembic==1.10.4
databases==0.7.0
fastapi==0.95.2
pydantic==1.10.7
sendgrid==6.10.0
SQLAlchemy==1.4.48
psycopg2-binary==2.9.6
uvicorn==0.22.0
asyncpg==0.24.0
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.7.0
requests==2.28.2
requests-html==0.10.0
requests-oauthlib==1.3.0
langchain==0.0.312
langchain-experimental==0.0.28
langchainhub==0.1.13
langcodes==3.3.0
python-jose==3.3.0